+++ src/Licensing/Models/LicenseKeyDataModel.cs
@@ -6,7 +6,7 @@ namespace Licensing.Models {
     public class LicenseKeyDataModel {
 
         public string Email { get; set; }
-        public int EpochTime { get; set; }
+        public float EpochTime { get; set; }
         public PermissionsModel Permissions { get; set; }
 
     }
